added embed collapses
